Changed Ratio to use 16-bit num/denom instead of 32 (saves a lot of boxed values)

+++ include/doremir/ratio.h
@@ -10,8 +10,8 @@
     @{
     */
 
-typedef int32_t doremir_ratio_num_t;
-typedef int32_t doremir_ratio_denom_t;
+typedef int16_t doremir_ratio_num_t;
+typedef int16_t doremir_ratio_denom_t;
 typedef struct _doremir_ratio_t * doremir_ratio_t;
 doremir_ratio_t doremir_ratio_create(doremir_ratio_num_t,
                                      doremir_ratio_denom_t);
